+++ src/main/java/de/maxhenkel/corpse/entities/DummyPlayer.java
@@ -23,7 +23,7 @@ public class DummyPlayer extends RemotePlayer {
                 setItemSlot(type, equipment.get(type.ordinal()));
             }
         }
-        AttributeInstance attribute = getAttributes().getInstance(NeoForgeMod.NAMETAG_DISTANCE.get());
+        AttributeInstance attribute = getAttributes().getInstance(NeoForgeMod.NAMETAG_DISTANCE.value());
         if (attribute != null) {
             attribute.setBaseValue(0D);
         }