+++ changelog.md
@@ -1,1 +1,2 @@
-- Ported to NeoForge
+- Updated NeoForge
+- Fixed corpse location misaligned