+++ changelog.md
@@ -1,1 +1,2 @@
 - Added mod icon
+- Updated corelib