+++ changelog.md
@@ -1,3 +1,1 @@
-- Added JEI to mods.toml
-- Added Jade to mods.toml
-- Switch to NeoGradle
+- Added mod icon