+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed container registration
+- Update to 1.21